packages/nex-invoice-worker/pyproject.toml
[project]
name = "nex-invoice-worker"
version = "1.0.0"
description = "Shared invoice worker package for multi-tenant Temporal workflows"
requires-python = ">=3.11"
dependencies = [
    "nex-config",
    "temporalio>=1.4.0",
    "pydantic-settings>=2.0.0",
    "zeep>=4.2.1",
    "httpx>=0.25.0",
    "asyncpg>=0.29.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
